% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDMC.R
\name{CellDMC}
\alias{CellDMC}
\title{A method that allows the identification of differentially methylated 
cell-types and their directionality of change}
\usage{
CellDMC(beta.m, pheno.v, frac.m, mode = c("improved", "basic"),
  adjPMethod = "fdr", adjPThresh = 0.05, DiffThresh = 0.1, mc.cores = 1)
}
\arguments{
\item{beta.m}{A beta value matrix with rows labeling the CpGs and columns labeling samples. 
This contains all CpGs, from which you want to find DMCTs.}

\item{pheno.v}{A vector of phenotype. Each item has to be 0 or 1, with 1 labeling case and 
0 labeling control.}

\item{frac.m}{A matrix contains fractions of each cell-type. Each row labels a sample, with 
the same order of the columns in beta.m. Each column labels a cell-type. 
Column names, which are names of cell-types, are required. The rowSums of 
frac.m should be 1, and all values should be greater than 0 and less than 1.}

\item{mode}{We provide two modes of CellDMC. One is 'basic' algorithm, and the other is '
improved' algorithm. \code{mode} can be either of 'improved' or 'basic', with 
default as 'improved'. For more details, pls refer to the reference.}

\item{adjPMethod}{A method to adjust p values. The method can be any of method accepted by 
\code{\link{p.adjust}}.}

\item{adjPThresh}{A numeric value, default as 0.05. This is used to call significant DMCTs. 
Adjusted p values less than this threshold will be picked.}

\item{DiffThresh}{A DNAm diff threshold. The default is 0.1. For each cell-type, CpGs with 
absolute DNAm change greater than this threshold will be treated as DMCTs. 
Pls note that this threshold is only used in 'improved' mode.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run 
simultaneously. The defatul is 1, which means no parallelization.}
}
\value{
A list with the following two items.

dmct
A matrix tells wheter the input CpGs are DMCTs and DMCs. The first column gives 
whether a CpG is DMC or not. If the CpG is called as DMC, the value will be 1,
otherwise it is 0. The following columns give DMCTs for each cell-types. If a 
CpG is DMCT, the value will be 1 (hypermethylated for case compared to control) 
or -1 (hypomethylated for case compared to control). Otherwise, the value is 0 
(non-DMCT). The rows of this matrix are ordered as the same as input 
\code{beta.m}.

coe
This list contains several dataframe, which corresponds to each cel-type in \code{frac.m}.
For \code{basic} mode, each dataframe only contains coefficients of DMCTs for 
each cell-type. For \code{improved} mode, each dataframe contains all CpGs in 
input \code{beta.m}. For both modes, each dataframe has been ranked with 
significant level of DMCTs of corresponding cell-type respectively. Pls note 
that the order of rows in these dataframe is different from the order of rows 
in \code{dmct} matrix. You can reorder them using the rownames. All dataframes 
contains ranks(\code{rank}), estimated DNAm differences(\code{Delta}), 
estimated T statistics(\code{Tstat}), raw P values(\code{rawP}), and multiple 
hypothesis corrected P values(\code{adjP}).
}
\description{
An outstanding challenge of Epigenome-Wide Association Studies performed in 
complex tissues is the identification of the specific cell-type(s) responsible
 for the observed differential methylation. CellDMC a novel statistical 
 algorithm, which is able to identify not only differentially methylated positions,
  but also the specific cell-type(s) driving the differential methylation.
}
\examples{
data(centEpiFibIC.m)
data(DummyBeta.m)
out.l <- epidish(DummyBeta.m, centEpiFibIC.m, method = 'RPC')
frac.m <- out.l$estF
pheno.v <- rep(c(0, 1), each = 5)
celldmc.o <- CellDMC(DummyBeta.m, pheno.v, frac.m) # Pls note this is faked beta value matrix


}
\references{
Zheng SC, Beck S, Teschendorff AE. 
\emph{Identification of differentially methylated cell-types in Epigenome-Wide 
Association Studies.} In preparation (2018).
}
