% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cent12CT450k.R
\docType{data}
\name{cent12CT450k.m}
\alias{cent12CT450k.m}
\title{Whole blood reference of 12 blood cell subtypes for 450k array}
\format{
A matrix with 600 rows and 12 columns
}
\usage{
data(cent12CT450k.m)
}
\description{
This reference is constructed using data from Salas et al.(2022). It contains
 the following 12 blood cell subtypes. This reference is constructed for 450k arrays.
 See \code{cent12CT.m} for EPIC arrays.
}
\details{
\itemize{
  \item CD4+ naive T-cells
  \item Basophil cells
  \item CD4+ memory T-cells
  \item Memory B-cells
  \item Naive B-cells
  \item Regulatory T-Cells
  \item CD8+ memory T-cells
  \item CD8+ naive T-cells
  \item Eosinophils
  \item NK-cells
  \item Neutrophils 
  \item Monocytes
}
}
\references{
Qi Luo, Varun B Dwaraka, Qingwen Chen, Huige Tong, Tianyu Zhu,Kirsten
Seale, Joseph M Raffaele, Shijie C Zheng, Tavis L Mendez, Yulu Chen,
Natalia Carreras, Sofina Begum, Kevin Mendez, Sarah Voisin, Nir Eynon,
Jessica A Lasky-Su, Ryan Smith, Andrew E Teschendorff. \emph{A
  meta-analysis of immune-cell fractions at high resolution reveals
  novel associations with common phenotypes and health outcomes.} Genome
Med. (2023) 15(1): 59.
doi:\href{https://doi.org/10.1186/s13073-023-01211-5}{10.1186/s13073-023-01211-5}.
}
\keyword{datasets}
