% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centEpiFibFatIC.m.R
\docType{data}
\name{centEpiFibFatIC.m}
\alias{centEpiFibFatIC.m}
\title{Reference for breast tissue}
\format{
A matrix with 491 rows and 4 columns
}
\usage{
data(centEpiFibFatIC.m)
}
\description{
This reference was designed for estimating fractions of epithelial 
cells, fibroblasts, fat cells and total immune cells in breast tissue.
}
\details{
\itemize{
  \item Epi
  \item Fib
  \item Fat
  \item IC
}
}
\references{
Zheng SC, Webster AP, Dong D, Feber A, Graham DG, Sullivan R, Jevons S, Lovat LB, 
Beck S, Widschwendter M, Teschendorff AE
\emph{A novel cell-type deconvolution algorithm reveals substantial contamination by immune cells in saliva, buccal and cervix.}
Epigenomics (2018) 10: 925-940.
doi:\href{https://doi.org/10.2217/epi-2018-0037}{
10.2217/epi-2018-0037}.
}
\keyword{datasets}
