% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centEpiFibIC.m.R
\docType{data}
\name{centEpiFibIC.m}
\alias{centEpiFibIC.m}
\title{Reference for epithelial cells, fibroblasts, and total immune cells}
\format{A matrix with 716 rows and 3 columns}
\usage{
data(centEpiFibIC.m)
}
\description{
Similar to \code{centDHSbloodDMC.m}. But this referene contains centroids for
epithelial cells, fibroblasts, and total immune cells.
}
\details{
\itemize{
  \item Epi
  \item Fib
  \item IC
}
}
\references{
Zheng SC, Webster AP, Dong D, Feber A, Graham DG, Sullivan R, Jevons S, Lovat LB, 
Beck S, Widschwendter M, Teschendorff AE
\emph{A novel cell-type deconvolution algorithm reveals substantial contamination by immune cells in saliva, buccal and cervix.}
Epigenomics (2018) In Press.
}
\keyword{datasets}
