% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cent12CT.R
\docType{data}
\name{centUniLIFE.m}
\alias{centUniLIFE.m}
\title{DNAm reference matrix for 19 immune cell-types for blood of any age}
\format{
A matrix with 1906 CpGs (rows) and 19 immune cell-types (columns)
}
\usage{
data(centUniLIFE.m)
}
\description{
This reference is constructed using data from Salas et al.(2022) and
several cord-blood DNAm datasets. It contains the following 19 immune cell
subtypes. This reference can be used for EPIC and 450k arrays.
}
\details{
\itemize{
  \item cord-blood granulocytes
  \item cord-blood monocytes
  \item cord-blood nRBCs
  \item cord-blood B-cells
  \item cord-blood NK-cells
  \item cord-blood CD8+ T-cells
  \item cord-blood CD4+ T-cells
  \item adult CD4+ naive T-cells
  \item adult Basophil cells
  \item adult CD4+ memory T-cells
  \item adult Memory B-cells
  \item adult Naive B-cells
  \item adult Regulatory T-Cells
  \item adult CD8+ memory T-cells
  \item adult CD8+ naive T-cells
  \item adult Eosinophils
  \item adult NK-cells
  \item adult Neutrophils 
  \item adult Monocytes
}
}
\references{
Xiaolong Guo et al \emph{Universal high resolution cell-type
  deconvolution of blood tissue from any age}. Submitted.
}
\keyword{datasets}
