% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyBeta.m.R
\docType{data}
\name{DummyBeta.m}
\alias{DummyBeta.m}
\title{Dummy beta value matrix}
\format{A matrix with 1000 rows and 2 columns}
\usage{
data(DummyBeta.m)
}
\description{
A dataset containing dummy whole blood beta value matrix
The dataset is part of the 450k GEO dataset GSE80559
The tissue type is whole blood
To reduce the data size, only 1000 probes are included
330 probes are overlapped with centDHSblood.m
You can get the whole beta value matrix by
\code{exprs(GEOquery::getGEO('GSE80559')[[1]])}
}
\details{
\itemize{
  \item beta value matrix of 1000 probes and 2 samples
}
}
\keyword{datasets}
