% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiuDataSub.m.R
\docType{data}
\name{LiuDataSub.m}
\alias{LiuDataSub.m}
\title{Whole blood example beta value matrix}
\format{
A matrix with 500 CpGs and 50 columns
}
\usage{
data(LiuDataSub.m)
}
\description{
This beta value matrix is a subset matrix of Liu et al data(GSE42861). Beta 
values of 326 CpGs in the \code{centDHSbloodDMC} reference matrix and other 
randomly chosen 174 CpGs are included for 50 randomly chosen samples.
}
\details{
\itemize{
  \item beta value matrix of 500 CpGs and 50 samples
}
}
\keyword{datasets}
