% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centDHSbloodDMC.m.R
\docType{data}
\name{centDHSbloodDMC.m}
\alias{centDHSbloodDMC.m}
\title{Whole blood reference of 333 tsDHS-DMCs and 8 blood cell subtypes}
\format{A matrix with 333 rows and 8 columns}
\usage{
data(centDHSbloodDMC.m)
}
\description{
A dataset containing wole blood reference of 333 tsDHS-DMCs 
and 8 blood cell subtypes:
}
\details{
\itemize{
  \item B-cells
  \item CD4+ T-cells
  \item CD8+ T-cells 
  \item NK-cells
  \item Monocytes
  \item Neutrophils 
  \item Eosinophils
  \item Granulocytes
}
}
\keyword{datasets}
