% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centEpiFibIC.m.R
\docType{data}
\name{centEpiFibIC.m}
\alias{centEpiFibIC.m}
\title{Reference for genenric epithelial tissue}
\format{
A matrix with 716 rows and 3 columns
}
\usage{
data(centEpiFibIC.m)
}
\description{
This reference could be used to estimate pproportions of epithelial cells, 
fibroblasts, and total immune cells in epithelial tissues.
}
\details{
\itemize{
  \item Epi
  \item Fib
  \item IC
}
}
\references{
Zheng SC, Webster AP, Dong D, Feber A, Graham DG, Sullivan R, Jevons S, Lovat LB, 
Beck S, Widschwendter M, Teschendorff AE
\emph{A novel cell-type deconvolution algorithm reveals substantial contamination by immune cells in saliva, buccal and cervix.}
Epigenomics (2018) 10: 925-940.
doi:\href{https://doi.org/10.2217/epi-2018-0037}{
10.2217/epi-2018-0037}.
}
\keyword{datasets}
