% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyBeta.m.R
\docType{data}
\name{DummyBeta.m}
\alias{DummyBeta.m}
\title{Dummy beta value matrix}
\format{
A matrix with 2000 CpGs and 10 columns
}
\usage{
data(DummyBeta.m)
}
\description{
A faked beta value matrix of 2000 CpGs and 10 samples. 
Only used for demostration purpose..
}
\details{
\itemize{
  \item beta value matrix of 2000 CpGs and 10 samples
}
}
\keyword{datasets}
