% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidish.R
\name{epidish}
\alias{epidish}
\title{Epigenetic Dissection of Intra-Sample-Heterogeneity}
\usage{
epidish(avdata.m, ref.m, method = c("RPC", "CBS", "CP"), maxit = 50,
  nu.v = c(0.25, 0.5, 0.75))
}
\arguments{
\item{avdata.m}{a data matrix with rows labeling the molecular features (should use same ID as in cent.m) and columns labeling samples (e.g. primary tumour specimens). No missing values are allowed and all values should be positive or zero. In the case of DNA methylation, these are beta-values.}

\item{ref.m}{a matrix of reference "centroids", i.e. representative molecular profiles, for a number of cell subtypes. rows label molecular features (e.g. CpGs,...) and columns label the cell-type. IDs need to be provided as rownames and colnames, respectively. No missing values are allowed, and all values in this matrix should be positive or zero. For DNAm data, values should be beta-values.}

\item{method}{chioce of a reference-based method ("RPC","CBS","CP")}
}
\value{
CP-mode
a list with the following entries: estF: the estimated cell fraction matrix; ref: the reference centroid matrix used; dataREF: the input data matrix over the probes defined in the reference matrix

CBS-mode
a list with the following entries: estF: the estimated cell fraction matrix; nu: a vector of "best" nu-parameter for each sample; 
ref: the reference centroid matrix used;
dataREF: the input data matrix over the probes defined in the reference matrix

RPC-mode
### a list with the following entries: estF: the estimated cell fraction matrix;
ref: the reference centroid matrix used; 
dataREF: the input data matrix over the probes defined in the reference matrix
}
\description{
a reference-based function to infer the proportions of a priori known cell subtypes present in a sample representing a mixture of such cell-types. Inference proceeds via one of 3 methods (Robust Partial Correlations-RPC, Cibersort (CBS), Constrained Projection (CP)), as determined by user.
}
\examples{
  #  library(EpiDISH)
  #  data(centDHSbloodDMC)
  #  out.l <- epidish(avdata.m, centDHSbloodDMC.m, method = "RPC") 
  #  estF.m <- out.l$estF
  ## avdata.m is from samples you would like to infer weights of cell subtypes
  ## estF.m is the inferred fractions
}

